pyproject.toml
[tool.poetry]
name = "spotify-playlist-curator-app"
version = "0.1.0"
description = ""
authors = ["mduruji <mczduruji@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
spotipy = "^2.23.0"
flask = "^3.0.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
